feat(APIM-562): rename module deal-folder to site-document
