+++ src/modules/site-document/interceptor/site-document-exception-transform.interceptor.ts
@@ -1,10 +1,10 @@
 import { BadRequestException, CallHandler, ExecutionContext, Injectable, NestInterceptor } from '@nestjs/common';
-import { UploadFileInDealFolderExistsException } from '@ukef/modules/deal-folder/exception/upload-file-in-deal-folder-exists.exception';
-import { UploadFileInDealFolderSiteNotFoundException } from '@ukef/modules/deal-folder/exception/upload-file-in-deal-folder-site-not-found.exception';
+import { UploadFileInDealFolderExistsException } from '@ukef/modules/site-document/exception/upload-file-in-deal-folder-exists.exception';
+import { UploadFileInDealFolderSiteNotFoundException } from '@ukef/modules/site-document/exception/upload-file-in-deal-folder-site-not-found.exception';
 import { catchError, Observable, throwError } from 'rxjs';
 
 @Injectable()
-export class DealFolderExceptionTransformInterceptor implements NestInterceptor {
+export class SiteDocumentExceptionTransformInterceptor implements NestInterceptor {
   intercept(_context: ExecutionContext, next: CallHandler): Observable<any> {
     return next.handle().pipe(
       catchError((err) =>
