+++ test/site-document/post-document-in-deal-folder.api-test.ts
@@ -2,8 +2,8 @@ import { RestError } from '@azure/storage-file-share';
 import { GraphError } from '@microsoft/microsoft-graph-client';
 import { BUYER_NAME, DTFS_MAX_FILE_SIZE_BYTES, ENUMS, FILE_LOCATION_PATH, SHAREPOINT } from '@ukef/constants';
 import { DocumentTypeEnum } from '@ukef/constants/enums/document-type';
-import { UploadFileInDealFolderParamsDto } from '@ukef/modules/deal-folder/dto/upload-file-in-deal-folder-params.dto';
-import { UploadFileInDealFolderRequestDto } from '@ukef/modules/deal-folder/dto/upload-file-in-deal-folder-request.dto';
+import { UploadFileInDealFolderParamsDto } from '@ukef/modules/site-document/dto/upload-file-in-deal-folder-params.dto';
+import { UploadFileInDealFolderRequestDto } from '@ukef/modules/site-document/dto/upload-file-in-deal-folder-request.dto';
 import { IncorrectAuthArg, withClientAuthenticationTests } from '@ukef-test/common-tests/client-authentication-api-tests';
 import { withStringFieldValidationApiTests } from '@ukef-test/common-tests/request-field-validation-api-tests/string-field-validation-api-tests';
 import { withDealIdParamValidationApiTests } from '@ukef-test/common-tests/request-param-validation-api-tests/deal-id-param-validation-api-tests';
